requirements.txt
Django>=5,<6
djangorestframework>=3.15,<4
stripe>=7,<8
python-dotenv>=1,<2